move header and list from components to containers, as they cant be tested individually
